feat(dragdrop): Extract the hit testing algorithm
